Scalable event tag index: sharded storage, streaming API, bucket-local chunks
Rewrites the event tag index for production scale (100M+ objects):

- Sharded inverted index with delta buffering and background compaction
- Chunks stored in user bucket (.minio.tag-index/) as text+zstd (~50% smaller)
- Only ~1 KB meta stays in .minio.sys per bucket regardless of object count
- Paginated query API (list-by-tag) and streaming API (stream-by-tag)
- 4 update workers with 100K channel capacity (was 1 worker, 10K cap)
- Non-blocking startup — loads only tiny meta files
- Auto-migration from all previous index formats (v0 JSON, v1 sharded JSON)
- Tests validating 50K-250K object chunks with compression benchmarks

Also updates console dependency and release workflow naming.

+++ cmd/admin-handlers.go
@@ -3594,9 +3594,9 @@ func (a adminAPIHandlers) RebuildTagIndexHandler(w http.ResponseWriter, r *http.
 	}
 
 	resp := struct {
-		Status string                    `json:"status"`
-		Bucket string                    `json:"bucket"`
-		Counts map[string]map[string]int `json:"counts"`
+		Status string                      `json:"status"`
+		Bucket string                      `json:"bucket"`
+		Counts map[string]map[string]int64 `json:"counts"`
 	}{
 		Status: "success",
 		Bucket: bucket,
